% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivolcanoonekey.R
\name{multivolcanoonekey}
\alias{multivolcanoonekey}
\title{multivolcanoonekey}
\usage{
multivolcanoonekey(
  res,
  cluster,
  logFC,
  padj,
  gene,
  top10 = F,
  title = NA,
  colors = NA,
  levels = NA
)
}
\arguments{
\item{res}{data.frame of DEGs to be plot}

\item{cluster}{the colname of clusters}

\item{logFC}{the colname of log2FoldChange}

\item{padj}{the colname of adjusted P value}

\item{gene}{the colname of gene}

\item{top10}{boolean if top10 DEGs labeled}

\item{title}{the title of the plot}

\item{colors}{the vector of colors of the clusters}

\item{levels}{the vector of levels of the clusters}
}
\value{
a ggplot2 object of multiple volcano plots
}
\description{
multivolcanoonekey
}
\examples{
library(ggonekey)
multivolcanoonekey(
    res = DEG,
    cluster = "cluster",
    logFC = "logFC",
    padj = "adj.P.Val",
    gene = "gene",
    top10 = T
)
}
