% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roconekey.R
\name{roconekey}
\alias{roconekey}
\title{roconekey}
\usage{
roconekey(
  data,
  start,
  end = start,
  title = NULL,
  y = 0.45,
  truth,
  nrow = 1,
  colors = c("red", "orange")
)
}
\arguments{
\item{data}{the data to be drawn}

\item{start}{the index where the columns to be drawn starts}

\item{end}{the index where the columns to be drawn ends(default end = start)}

\item{title}{the title of the plot}

\item{y}{how high the auc label should be}

\item{truth}{the colname which contains the ground truth}

\item{nrow}{how many rows of the plots will be}

\item{colors}{a vector of two colors of the curves}
}
\value{
a ggplot2 object of multiple ROC plots
}
\description{
roconekey
}
\examples{
library(ggplot2)
library(multipleROC)
roconekey(data = mtcars, start = 2, end = 7, truth = "am", nrow = 2)
}
