% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeyonekey.R
\name{sankeyonekey}
\alias{sankeyonekey}
\title{sankeyonekey}
\usage{
sankeyonekey(make_long, title = NULL)
}
\arguments{
\item{make_long}{the data.frame after ggsankey::make_long()}

\item{title}{the title of the plot}
}
\value{
a ggplot2 object of a sankey plot
}
\description{
sankeyonekey
}
\examples{
library(ggplot2)
library(ggsankey)
data <- data.frame(
  group1 = c(1, 1, 2, 2),
  group2 = c(3, 4, 3, 4)
)
data <- ggsankey::make_long(data, group1, group2)

data$node <- factor(data$node, levels = rev(unique(data$node)))
data$next_node <- factor(data$next_node, levels = rev(unique(data$next_node)))

sankeyonekey(make_long = data, title = "TITLE")
}
