% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoonekey.R
\name{volcanoonekey}
\alias{volcanoonekey}
\title{Title}
\usage{
volcanoonekey(
  data,
  gene,
  logfc,
  pvalue,
  title = NULL,
  labeled = F,
  colors = c("blue", "gray", "red")
)
}
\arguments{
\item{data}{the data.frame to be drawn}

\item{gene}{the colname of genes}

\item{logfc}{the colname of log2FoldChange}

\item{pvalue}{the colname of p value or adjusted p value}

\item{title}{the title of the plot}

\item{labeled}{if the top 10 genes are to be labeled}

\item{colors}{a vector of three colors on behalf of "down", "ns" and "up"}
}
\value{
a ggplot2 object of a volcano plot
}
\description{
Title
}
\examples{
library(DESeq2)
library(ggplot2)
library(ggrepel)
set.seed(123)
cnts <- matrix(rnbinom(n = 1000, mu = 100, size = 1 / 0.5), ncol = 10)
cond <- factor(rep(1:2, each = 5))

dds <- DESeq2::DESeqDataSetFromMatrix(cnts, DataFrame(cond), ~cond)

dds <- DESeq2::DESeq(dds)
res <- DESeq2::results(dds)

DEG <- as.data.frame(res)
DEG$gene <- rownames(DEG)

volcanoonekey(
  data = DEG,
  gene = "gene",
  logfc = "log2FoldChange",
  pvalue = "padj",
  title = "TITLE",
  labeled = TRUE
)
}
